+++ server/modules/explorer.js
@@ -106,7 +106,7 @@ const processWinningBlocks = async () => {
                     )
                 ) {
                     const lastBlockData = blocks.data.blocks[0]
-                    const message = `We got last block at ${timestamp("-", lastBlockData.dateTime)} on height ${lastBlockData.blockHeight} with coinbase ${lastBlockData.transactions.coinbase / 10 ** 9} mina.`
+                    const message = `We got last block at \`${timestamp("-", lastBlockData.dateTime)}\` on height \`${lastBlockData.blockHeight}\` with coinbase \`${lastBlockData.transactions.coinbase / 10 ** 9}\` mina.`
                     sendMessage('REWARDS', message)
                 }
             } catch (e) {}