+++ server/modules/hello.js
@@ -1,10 +1,9 @@
-const {hostname} = require("os")
 const {sendAlert} = require("../helpers/messangers")
 
 const processHello = () => {
-    const message = `Hello from Mina Monitor!`
+    const message = `Mina Monitor started successful! Have nice work!`
 
-    sendAlert("HELLO", message)
+    sendAlert("HELLO", message, false)
 }
 
 module.exports = {