+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 + [x] Server: added config options for `discord`: `webhook`, `webhookInfo`, `webhookAlert`, `botName` 
 + [x] Server: to config parameters `alertToTelegram`, `alertToDiscrod` added value `PRICE` 
 + [x] Server: to config parameters `alertToTelegram`, `alertToDiscrod` added value `REWARDS`
++ [x] Server: to config parameters `alertToTelegram`, `alertToDiscrod` added value `UPTIME`
 
 ### 2.0.3
 + [x] Server: added support for time interval string values for parameter `restartAfterNotSynced`